Altered kinematic_wave_ssf
Based on a discussion between myself and Willem v. Verseveld, we decided there was an error in the previous accounting of the recharge rate, their was a type to represent the lateral flow rate of the previous timestep.
Furthermore, it was decided to make use of the waterbalance approach to estimate the groundwater level zi, instead of using its relation with respect to lateral flow ssf_n. The new estimate is more representative for thw whole grid cell (instead of its boundary only and the Newton Raphson procedure iterates faster to a stable solution.